test: extract testcontainers setup
