+++ backend/src/test/kotlin/com/example/codex/AbstractIntegrationTest.kt
@@ -1,0 +1,29 @@
+package com.example.codex
+
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.DynamicPropertyRegistry
+import org.springframework.test.context.DynamicPropertySource
+import org.testcontainers.containers.PostgreSQLContainer
+import org.testcontainers.junit.jupiter.Container
+import org.testcontainers.junit.jupiter.Testcontainers
+
+@Testcontainers
+@SpringBootTest
+abstract class AbstractIntegrationTest {
+    companion object {
+        @Container
+        private val postgres = PostgreSQLContainer<Nothing>("postgres:16-alpine").apply {
+            withDatabaseName("codex")
+            withUsername("codex")
+            withPassword("codex")
+        }
+
+        @JvmStatic
+        @DynamicPropertySource
+        fun datasourceConfig(registry: DynamicPropertyRegistry) {
+            registry.add("spring.datasource.url", postgres::getJdbcUrl)
+            registry.add("spring.datasource.username", postgres::getUsername)
+            registry.add("spring.datasource.password", postgres::getPassword)
+        }
+    }
+}
