+++ backend/src/test/kotlin/com/example/codex/repository/UserRepositoryIT.kt
@@ -1,37 +1,15 @@
 package com.example.codex.repository
 
+import com.example.codex.AbstractIntegrationTest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.DynamicPropertyRegistry
-import org.springframework.test.context.DynamicPropertySource
-import org.testcontainers.containers.PostgreSQLContainer
-import org.testcontainers.junit.jupiter.Container
-import org.testcontainers.junit.jupiter.Testcontainers
 
-@Testcontainers
-@SpringBootTest
 class UserRepositoryIT @Autowired constructor(
     private val userRepository: UserRepository,
-) {
-    companion object {
-        @Container
-        private val postgres = PostgreSQLContainer<Nothing>("postgres:16-alpine").apply {
-            withDatabaseName("codex")
-            withUsername("codex")
-            withPassword("codex")
-        }
 
-        @JvmStatic
-        @DynamicPropertySource
-        fun datasourceConfig(registry: DynamicPropertyRegistry) {
-            registry.add("spring.datasource.url") { postgres.jdbcUrl }
-            registry.add("spring.datasource.username") { postgres.username }
-            registry.add("spring.datasource.password") { postgres.password }
-        }
-    }
+) : AbstractIntegrationTest() {
 
     @Test
     fun `saves and retrieves user by username`() {
